requirements.txt
flask==2.0.1
gunicorn==20.1.0
python-dotenv==0.19.0
PyPDF2==2.10.5
requests==2.26.0
scikit-learn==0.24.2
numpy==1.21.2
Pillow==8.3.2
fpdf==1.7.2
google-cloud-vision==2.6.3
pdf2image==1.16.0
flask-cors==3.0.10 